[MIG] shopinvader_api_wishlist
