+++ shopinvader_api_wishlist/__init__.py
@@ -1,4 +1,3 @@
 from . import wizard
 from . import schemas
-from . import dependencies
 from . import routers
