+++ shopinvader_api_wishlist/__manifest__.py
@@ -4,27 +4,24 @@
 
 {
     "name": "Shopinvader Wishlist",
-    "summary": """
-        Handle shop wishlist""",
-    "version": "16.0.1.0.2",
+    "summary": "Handle shop wishlist",
+    "version": "18.0.1.0.0",
     "license": "AGPL-3",
     "author": "ACSONE SA/NV, Camptocamp,Odoo Community Association (OCA)",
     "website": "https://github.com/shopinvader/odoo-shopinvader",
     "depends": [
-        "fastapi",
         "sale_wishlist",
         "extendable_fastapi",
+        "shopinvader_router_helper",
         "shopinvader_api_security_sale",
         "shopinvader_schema_sale",
         "shopinvader_sale_cart",
-        "shopinvader_filtered_model",
     ],
-    "demo": [],
     "data": [
         "security/groups.xml",
         "security/acl_sale_product_set_wizard.xml",
         "security/rule+acl_product_set.xml",
         "security/rule+acl_product_set_line.xml",
     ],
-    'installable': False,
+    "installable": True,
 }
