+++ shopinvader_api_wishlist/readme/CONTRIBUTORS.md
@@ -1,0 +1,2 @@
+- Laurent Mignon \<laurent.mignon@acsone.com\>
+- Simone Orsi \<simone.orsi@camptocamp.com\>
