--- shopinvader_api_wishlist/readme/CONTRIBUTORS.rst
@@ -1,2 +1,0 @@
-* Laurent Mignon <laurent.mignon@acsone.com>
-* Simone Orsi <simone.orsi@camptocamp.com>
