+++ shopinvader_api_wishlist/readme/CREDITS.md
@@ -1,5 +1,5 @@
 The development of this module has been financially supported by:
 
-* Camptocamp
-* Cosanum
-* Alcyon Belux
+- Camptocamp
+- Cosanum
+- Alcyon Belux
