+++ shopinvader_api_wishlist/security/groups.xml
@@ -2,14 +2,13 @@
 <!-- Copyright 2023 ACSONE SA/NV
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
 <odoo>
-  <record id="shopinvader_wishlist_user_group" model="res.groups">
-    <field name="name">Shopinvader Wishlist user</field>
-      <field
+    <record id="shopinvader_wishlist_user_group" model="res.groups">
+        <field name="name">Shopinvader Wishlist user</field>
+        <field
             name="implied_ids"
             eval="[
       (4, ref('fastapi.group_fastapi_endpoint_runner')),
       ]"
         />
-  </record>
-
+    </record>
 </odoo>
