+++ shopinvader_api_wishlist/security/rule+acl_product_set.xml
@@ -2,7 +2,6 @@
 <!-- Copyright 2023 ACSONE SA/NV
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
 <odoo>
-
     <record model="ir.model.access" id="shopinvader_product_set_access">
         <field name="name">product.set shopinvader user full access</field>
         <field name="model_id" ref="product_set.model_product_set" />
@@ -11,19 +10,22 @@
         <field name="perm_create" eval="1" />
         <field name="perm_unlink" eval="1" />
         <field name="perm_write" eval="1" />
-  </record>
+    </record>
 
-  <record model="ir.rule" id="shopinvader_user_access_own_product_set_rule">
-    <field name="name">shopinvader user can only access its own product set</field>
-    <field name="model_id" ref="product_set.model_product_set" />
-    <field name="domain_force">[
+    <record model="ir.rule" id="shopinvader_user_access_own_product_set_rule">
+        <field name="name">shopinvader user can only access its own product set</field>
+        <field name="model_id" ref="product_set.model_product_set" />
+        <field name="domain_force">[
       ('partner_id', '=', authenticated_partner_id)
         ]
     </field>
-    <field name="groups" eval="[(6, 0, [ref('shopinvader_wishlist_user_group')])]" />
-    <field name="perm_read" eval="1" />
-    <field name="perm_create" eval="1" />
-    <field name="perm_write" eval="1" />
-    <field name="perm_unlink" eval="0" />
-  </record>
+        <field
+            name="groups"
+            eval="[(6, 0, [ref('shopinvader_wishlist_user_group')])]"
+        />
+        <field name="perm_read" eval="1" />
+        <field name="perm_create" eval="1" />
+        <field name="perm_write" eval="1" />
+        <field name="perm_unlink" eval="0" />
+    </record>
 </odoo>
