+++ shopinvader_api_wishlist/security/rule+acl_product_set_line.xml
@@ -2,7 +2,6 @@
 <!-- Copyright 2023 ACSONE SA/NV
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
 <odoo>
-
     <record model="ir.model.access" id="shopinvader_product_set_line_access">
         <field name="name">product.set.line shopinvader user full access</field>
         <field name="model_id" ref="product_set.model_product_set_line" />
@@ -11,20 +10,23 @@
         <field name="perm_create" eval="1" />
         <field name="perm_unlink" eval="1" />
         <field name="perm_write" eval="1" />
-  </record>
-  <record model="ir.rule" id="shopinvader_user_access_own_product_set_line_rule">
-    <field
+    </record>
+    <record model="ir.rule" id="shopinvader_user_access_own_product_set_line_rule">
+        <field
             name="name"
         >shopinvader user can only access its own product set lines</field>
-    <field name="model_id" ref="product_set.model_product_set_line" />
-    <field name="domain_force">[
+        <field name="model_id" ref="product_set.model_product_set_line" />
+        <field name="domain_force">[
       ('product_set_id.partner_id', '=', authenticated_partner_id)
         ]
     </field>
-    <field name="groups" eval="[(6, 0, [ref('shopinvader_wishlist_user_group')])]" />
-    <field name="perm_read" eval="1" />
-    <field name="perm_create" eval="1" />
-    <field name="perm_write" eval="1" />
-    <field name="perm_unlink" eval="0" />
-  </record>
+        <field
+            name="groups"
+            eval="[(6, 0, [ref('shopinvader_wishlist_user_group')])]"
+        />
+        <field name="perm_read" eval="1" />
+        <field name="perm_create" eval="1" />
+        <field name="perm_write" eval="1" />
+        <field name="perm_unlink" eval="0" />
+    </record>
 </odoo>
